% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend}
\alias{backend}
\alias{backend_set}
\alias{backend_get}
\title{Set location of data acquisition.}
\usage{
backend_set(itis = "api", col = itis, ncbi = itis, theplantlist = itis,
  path = "~/.taxize_local", itis_user = NULL, itis_pwd = NULL,
  col_user = NULL, col_pwd = NULL)

backend_get()
}
\arguments{
\item{itis, col, ncbi, theplantlist}{(character) The backend to query data
from. Defaults to \code{api}, which means we query resources on the web.
Alternatively, use \code{local}, which uses local versions of databases on
your own machine. This option does some checks to make sure things are
setup correctly.}

\item{path}{The path to store local databases. Individual databases are
within this base path.}

\item{itis_user}{ITIS user name}

\item{itis_pwd}{ITIS password}

\item{col_user}{Catalogue of Life user name}

\item{col_pwd}{Catalogue of Life password}
}
\description{
Set location of data acquisition.
}
\details{
All other data source parameters in \code{\link[=backend_set]{backend_set()}} inherit
from the first \code{itis}, which is by default set to \code{itis = "api"}

Note that the only option for Theplantlist data is "local". Alternatively,
you can download raw csv files of their data using
\code{\link[taxize:tpl_get]{taxize::tpl_get()}}, or use the \pkg{Taxonstand} package that
downloads csv files and uses regex locally in R.
}
\examples{
\dontrun{
# set all to remote api
backend_set("api")
backend_get()

# set all to local
backend_set("local")
backend_get()

# set individual data sources to different settings
backend_set(col = "local", ncbi = "local")
backend_get()

# Set username and password for COL connection
backend_set(col_user="root")
}
}
