% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_children.R
\name{children}
\alias{children}
\title{Retrieve immediate descendents of a taxon}
\usage{
children(x, db = "ncbi", verbose = TRUE, ...)
}
\arguments{
\item{x}{Vector of taxon keys for the given database}

\item{db}{The database to search}

\item{verbose}{Print verbose messages}

\item{...}{Additional arguments passed to database specific function.}
}
\value{
list of data.frames with the columns: childtaxa_id,  childtaxa_name,
childtaxa_rank. This is exactly equivalent to the output of
'taxize::children.
}
\description{
Retrieve immediate descendents of a taxon
}
\examples{
\dontrun{
children(c(3700, 2))
}
}
