% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Retrieve the taxonomic hierarchies from local database}
\usage{
classification(x, db = "ncbi", verbose = TRUE, ...)
}
\arguments{
\item{x}{Vector of taxon keys for the given database}

\item{db}{The database to search}

\item{verbose}{Print verbose messages}

\item{...}{Additional arguments passed to database specific classification functions.}
}
\value{
list of data.frames with the columns: name, rank, and id. This is
exactly equivalent to the output of 'taxize::classification'.
}
\description{
This function is equivalent to the taxize::classification function, except
that it uses a local database (so is much faster) and is currently
restricted to handling NCBI taxa. The output is identical to
taxize::classification.
}
\examples{
\dontrun{
classification(c(3702, 9606))
}
}
