% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db}
\alias{db}
\alias{db_col}
\alias{db_itis}
\alias{db_plantlist}
\title{Download taxonomic databases}
\usage{
db_itis(verbose = TRUE, user = "sacmac", pwd = NULL)

db_plantlist(user = NULL, pwd = NULL, verbose = TRUE)

db_col(user = "root", pwd = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{(logical) Print messages. Default: \code{TRUE}}

\item{user}{User name}

\item{pwd}{Password}
}
\value{
Downloads sql database, loads it into PostgreSQL or MySQL,
cleans up unneeded files, returns path of sql DB
}
\description{
Download taxonomic databases
}
\details{
Supported:
\itemize{
\item ITIS
\item the PlantList
\item Catalogue of Life
}
}
\examples{
\dontrun{
db_itis()
db_col()
db_plantlist()
}
}

