% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_download.R
\name{db_download}
\alias{db_download}
\alias{db_download_ncbi}
\alias{db_download_itis}
\alias{db_download_tpl}
\alias{db_download_col}
\alias{db_download_gbif}
\alias{db_download_wikidata}
\title{Download taxonomic databases}
\usage{
db_download_ncbi(verbose = TRUE)

db_download_itis(verbose = TRUE)

db_download_tpl(verbose = TRUE)

db_download_col(verbose = TRUE)

db_download_gbif(verbose = TRUE)

db_download_wikidata(verbose = TRUE)
}
\arguments{
\item{verbose}{(logical) Print messages. Default: \code{TRUE}}
}
\value{
Path to the downloaded SQL database
}
\description{
Download taxonomic databases
}
\details{
Downloads sql database, cleans up unneeded files, returns path
to sql file
}
\section{Supported}{

\itemize{
\item ITIS - SQLite
\item The PlantList - SQLite
\item Catalogue of Life - MariaDB
\item GBIF - SQLite
\item Wikidata - SQLite
}
}

\section{Links}{

\itemize{
\item ITIS - xx
\item The PlantList - xx
\item Catalogue of Life -
via http://www.catalogueoflife.org/content/annual-checklist-archive
\item GBIF - xx
\item Wikidata - via https://zenodo.org/record/1213477
}
}

\section{Beware}{

COL database loading takes a long time, e.g., 30 minutes. you may
want to run it in a separate R session, or just look at the db_load_col fxn
and run the commands in your shell.
}

\examples{
\dontrun{
# ITIS
# db_download_itis()
# src_itis()

# Plantlist
# db_download_tpl()
# src_tpl()

# COL
# db_download_col()
# src_col()

# GBIF
# db_download_gbif()
# src_gbif()

# NCBI
# db_download_ncbi()
# src_ncbi()

# Wikidata
# db_download_wikidata()
# src_wikidata()
}
}
\seealso{
\link{tdb_cache}
}
