% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_name2taxid.R
\name{name2taxid}
\alias{name2taxid}
\title{Convert species names to taxon IDs}
\usage{
name2taxid(x, db = "ncbi", verbose = TRUE, out_type = c("uid", "summary"),
  ...)
}
\arguments{
\item{x}{Vector of taxon keys for the given database}

\item{db}{The database to search}

\item{verbose}{Print verbose messages}

\item{...}{Additional arguments passed to database specific classification functions.}
}
\description{
\code{name2taxid} returns a vector and dies if there are any ambiguous
names. \code{name2taxid_map} returns a data.frame mapping names to ids.
}
\examples{
\dontrun{
name2taxid(c('Arabidopsis thaliana', 'pig'))
}
}
