% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{plantlist_SQL}
\alias{plantlist_SQL}
\title{ThePlantList sql initiation and query}
\usage{
plantlist_SQL(query, user = NULL, password = NULL, ...)
}
\description{
ThePlantList sql initiation and query
}
\keyword{internal}

