% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_collect}
\alias{sql_collect}
\title{Query and get data back into a data.frame}
\usage{
sql_collect(src, query, ...)
}
\arguments{
\item{src}{An \code{src} object, result of calling \code{\link{src_itis}},
\code{\link{src_col}}, or \code{\link{src_tpl}}}

\item{query}{A SQL query}

\item{...}{further args passed on to \code{\link[dplyr]{tbl}}}
}
\description{
Query and get data back into a data.frame
}
\details{
we run \code{\link[dplyr]{tbl}}, then \code{\link[dplyr]{collect}}
}
\examples{
\dontrun{
src <- src_itis()

# use SQL syntax
#sql_collect(src, "<some query>")
sql_collect(src, "select * from hierarchy limit 5")
## or pipe the src to sql_collect
src \%>\% sql_collect("select * from hierarchy limit 5")

# use dplyr verbs
src \%>\%
  tbl("hierarchy") \%>\%
  top_n(10)

## or create tbl object for repeated use
hiers <- src \%>\% tbl("hierarchy")
hiers \%>\% top_n(10)
hiers \%>\% select(tsn, level)
}
}

