% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_taxizedb}
\alias{src_col}
\alias{src_itis}
\alias{src_taxizedb}
\alias{src_tpl}
\title{ITIS src}
\usage{
src_itis(user = NULL, password = NULL, dbname = "ITIS", ...)

src_col(user = "root", password = NULL, dbname = "col", ...)

src_tpl(user = NULL, password = NULL, dbname = "plantlistdb", ...)
}
\arguments{
\item{user}{(character) user name}

\item{password}{(character) password}

\item{dbname}{(character) database name. Defaults: ITIS, col, and plantlistdb
for ITIS, COL, and ThePlantlist, respectively}

\item{...}{Further args passed on to \code{\link[dplyr]{src_postgres}} or
\code{\link[dplyr]{src_mysql}}}
}
\value{
an src object
}
\description{
ITIS src
}
\examples{
\dontrun{
src_itis()
src_col()
src_tpl()
}
}

