% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_taxid2name.R
\name{taxid2name}
\alias{taxid2name}
\title{Convert taxon IDs to scientific names}
\usage{
taxid2name(x, db = "ncbi", verbose = TRUE, ...)
}
\arguments{
\item{x}{Vector of taxon keys for the given database}

\item{db}{The database to search}

\item{verbose}{Print verbose messages}

\item{...}{Additional arguments passed to database specific classification functions.}
}
\value{
character vector of scientific names
}
\description{
Convert taxon IDs to scientific names
}
\examples{
\dontrun{
taxid2name(c(3702, 9606))
}
}
