% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_taxid2rank.R
\name{taxid2rank}
\alias{taxid2rank}
\title{Convert taxon IDs to scientific ranks}
\usage{
taxid2rank(x, db = "ncbi", verbose = TRUE, warn = TRUE, ...)
}
\arguments{
\item{x}{Vector of taxon keys (name or id) for the given database}

\item{db}{The database to search}

\item{verbose}{Print verbose messages}

\item{warn}{If TRUE, raise a warning if any taxon IDs can not be found}

\item{...}{Additional arguments passed to database specific classification functions.}
}
\value{
character vector of ranks
}
\description{
Convert taxon IDs to scientific ranks
}
\examples{
\dontrun{
taxid2rank(c(3701, 9606))
}
}
