% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxizedb-package.R
\docType{package}
\name{taxizedb-package}
\alias{taxizedb}
\alias{taxizedb-package}
\title{Taxonomic databases interface}
\description{
Taxonomic databases interface
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

