% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification}
\alias{classification}
\title{Retrieve the taxonomic hierarchies from local database}
\usage{
classification(src, taxa, db = "ncbi")
}
\arguments{
\item{src}{src_dbi database link to a local database}

\item{taxa}{Vector of taxonomy ids}

\item{db}{The database to search (currently only NCBI is implemented)}
}
\value{
list of data.frames with the columns: name, rank, and id. This is
exactly equivalent to the output of 'taxize::classification'.
}
\description{
This function is equivalent to the taxize::classification function, except
that it uses a local database (so is much faster) and is currently
restricted to handling NCBI taxa. The output is identical to
taxize::classification.
}
