% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{db}
\alias{db}
\alias{db_download_col}
\alias{db_download_itis}
\alias{db_download_tpl}
\alias{db_load_col}
\alias{db_load_itis}
\alias{db_load_tpl}
\title{Download and load taxonomic databases}
\usage{
db_download_itis(verbose = TRUE)

db_load_itis(path, user, pwd = NULL, verbose = TRUE)

db_download_tpl(verbose = TRUE)

db_load_tpl(path, user = NULL, pwd = NULL, verbose = TRUE)

db_download_col(verbose = TRUE)

db_load_col(path, user = "root", pwd = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{(logical) Print messages. Default: \code{TRUE}}

\item{path}{(character) path to the \code{.sql} database file}

\item{user}{(character) User name}

\item{pwd}{(character) Password, if any}
}
\value{
Downloads sql database, loads it into PostgreSQL or MySQL,
cleans up unneeded files, returns path of sql DB
}
\description{
Download and load taxonomic databases
}
\details{
Supported:
\itemize{
 \item ITIS - PostgreSQL
 \item the PlantList - PostgreSQL
 \item Catalogue of Life - MySQL
}

\code{db_download*()} functions are for downloading SQL DBs, and they return
the file path, but they don't load the database

\code{db_load*()} functions are for loading SQL DBs into the respective driver,
and they return the file path, but they don't load the database

Beware: COL database loading takes a long time, e.g., 30 minutes. you may
want to run it in a separate R session, or just look at the db_load_col fxn
and run the commands in your shell.
}
\examples{
\dontrun{
# ITIS
#db_download_itis() \%>\% db_load_itis()
x <- db_download_itis()
db_load_itis(x)

# the plant list
#db_download_tpl() \%>\% db_load_tpl()
x <- db_download_tpl()
db_load_tpl(x)

# catalogue of life
#db_download_col() \%>\% db_load_col()
x <- db_download_col()
db_load_col(x)
}
}

