% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_download.R
\name{db_download}
\alias{db_download}
\alias{db_download_ncbi}
\alias{db_download_itis}
\alias{db_download_tpl}
\alias{db_download_col}
\alias{db_download_gbif}
\alias{db_download_wikidata}
\title{Download taxonomic databases}
\usage{
db_download_ncbi(verbose = TRUE)

db_download_itis(verbose = TRUE)

db_download_tpl(verbose = TRUE)

db_download_col(verbose = TRUE)

db_download_gbif(verbose = TRUE)

db_download_wikidata(verbose = TRUE)
}
\arguments{
\item{verbose}{(logical) Print messages. Default: \code{TRUE}}
}
\value{
(character) path to the downloaded SQL database
}
\description{
Download taxonomic databases
}
\details{
Downloads sql database, cleans up unneeded files, returns path
to sql file
}
\section{Supported}{

All are using SQLite as the database
\itemize{
\item NCBI
\item ITIS
\item The PlantList
\item Catalogue of Life
\item GBIF
\item Wikidata
}
}

\section{Links}{

\itemize{
\item NCBI: ftp://ftp.ncbi.nih.gov/pub/taxonomy/
\item ITIS: https://www.itis.gov/downloads/index.html
\item The PlantList - http://www.theplantlist.org/
\item Catalogue of Life:
via http://www.catalogueoflife.org/content/annual-checklist-archive
\item GBIF: http://rs.gbif.org/datasets/backbone/
\item Wikidata: https://zenodo.org/record/1213477
}
}

\examples{
\dontrun{
# ITIS
# db_download_itis()
# src_itis()

# Plantlist
# db_download_tpl()
# src_tpl()

# COL
# db_download_col()
# src_col()

# GBIF
# db_download_gbif()
# src_gbif()

# NCBI
# db_download_ncbi()
# src_ncbi()

# Wikidata
# db_download_wikidata()
# src_wikidata()
}
}
\seealso{
\link{tdb_cache}
}
