% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_taxizedb}
\alias{src_taxizedb}
\alias{src_itis}
\alias{src_tpl}
\alias{src_col}
\alias{src_gbif}
\title{src - dplyr src objects}
\usage{
src_itis(user = NULL, password = NULL, dbname = "ITIS", ...)

src_tpl(user = NULL, password = NULL, dbname = "plantlist", ...)

src_col(user = "root", password = NULL, dbname = "col", ...)

src_gbif(path)
}
\arguments{
\item{user}{(character) user name}

\item{password}{(character) password}

\item{dbname}{(character) database name. Defaults: ITIS, col, and
plantlistdb for ITIS, COL, and ThePlantlist, respectively. GBIF uses
SQLite so doesn't have a database name}

\item{...}{Further args passed on to \code{\link[dplyr:src_postgres]{dplyr::src_postgres()}} or
\code{\link[dplyr:src_mysql]{dplyr::src_mysql()}}, or \code{\link[dplyr:src_sqlite]{dplyr::src_sqlite()}}}

\item{path}{(character) path to SQLite database}
}
\value{
an src object
}
\description{
src - dplyr src objects
}
\examples{
\dontrun{
# src_itis()
# src_tpl()
# src_col()
# src_gbif()
}
}
