% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_taxizedb}
\alias{src_taxizedb}
\alias{src_itis}
\alias{src_tpl}
\alias{src_col}
\alias{src_gbif}
\alias{src_ncbi}
\alias{src_wikidata}
\alias{src_wfo}
\title{src - dplyr src objects}
\usage{
src_itis(path = db_path("itis"), ...)

src_tpl(path = db_path("tpl"), ...)

src_col(path = db_path("col"), ...)

src_gbif(path = db_path("gbif"), ...)

src_ncbi(path = db_path("ncbi"), ...)

src_wikidata(path = db_path("wikidata"), ...)

src_wfo(path = db_path("wfo"), ...)
}
\arguments{
\item{path}{(character) path to SQLite database. by default
we use the function \code{\link[=db_path]{db_path()}} to get the path}

\item{...}{Further args passed on to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
an src object
}
\description{
src - dplyr src objects
}
\examples{
\dontrun{
# src_itis()
# src_tpl()
# src_col()
# src_gbif()
# src_ncbi()
# src_wikidata()
# src_wfo()
}
}
