% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_taxid2name.R
\name{taxid2name}
\alias{taxid2name}
\title{Convert taxon IDs to scientific names}
\usage{
taxid2name(x, db = "ncbi", verbose = TRUE, warn = TRUE, ...)
}
\arguments{
\item{x}{(character) Vector of taxon keys for the given database}

\item{db}{(character) The database to search}

\item{verbose}{(logical) Print verbose messages}

\item{warn}{(logical) If \code{TRUE}, raise a warning if any taxon IDs can not
be found}

\item{...}{Additional arguments passed to database specific classification
functions}
}
\value{
character vector of scientific names
}
\description{
Convert taxon IDs to scientific names
}
\examples{
\dontrun{
taxid2name(c(3702, 9606))
}
}
