% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxizedb-package.R
\docType{package}
\name{taxizedb-package}
\alias{taxizedb-package}
\alias{taxizedb}
\title{Taxonomic databases interface}
\description{
Taxonomic databases interface
}
\examples{
\dontrun{
library(dplyr)

# data source: ITIS
## download ITIS database
db_download_itis()
## connect to the ITIS database
src <- src_itis()
## use SQL syntax
sql_collect(src, "select * from hierarchy limit 5")
### or pipe the src to sql_collect
src \%>\% sql_collect("select * from hierarchy limit 5")
## use dplyr verbs
src \%>\%
  tbl("hierarchy") \%>\%
  filter(ChildrenCount > 1000)
## or create tbl object for repeated use
hiers <- src \%>\% tbl("hierarchy")
hiers \%>\% select(TSN, level)

# data source: The PlantList
## download tpl datababase
db_download_tpl()
## connecto the tpl database
src <- src_tpl()
## do queries
tpl <- tbl(src, "tpl")
filter(tpl, Family == "Pinaceae")

# data source: catalogue of life
## download col datababase
db_download_col()
## connec to the col database
src <- src_col()
## do queries
names <- tbl(src, "taxa")
select(names, taxonID, scientificName)

# data source: GBIF
## download gbif datababase
db_download_gbif()
## connecto the gbif database
src <- src_gbif()
## do queries
df <- tbl(src, "gbif")
select(df, taxonID, scientificName)
}
}
\keyword{package}
